final_script_integration_example.py
"""
Example of how to integrate Pinecone functionality into final_script.py.

This is not a complete implementation, but shows the key parts that need to be added
to integrate the Pinecone indexing and querying functionality.
"""

import streamlit as st
import os
import sys
from pathlib import Path

# Import the Pinecone integration module
import pinecone_integration

# Add this to your final_script.py imports
# from pinecone_integration import integrate_pinecone_workflow

def main():
    st.title("Research Paper Finder with Pinecone Integration")
    
    # Initialize session state variables if they don't exist
    if "chat_id" not in st.session_state:
        st.session_state.chat_id = None
    
    if "processed_data_folder" not in st.session_state:
        st.session_state.processed_data_folder = None
    
    # Integrate Pinecone workflow
    # This adds the necessary functions to this module
    pinecone_integration.integrate_pinecone_workflow(sys.modules[__name__])
    
    # Create tabs for different functionality
    tab1, tab2, tab3, tab4 = st.tabs(["Search Papers", "Process PDFs", "Index in Pinecone", "Query Index"])
    
    with tab1:
        st.header("Search and Download Papers")
        
        # Your existing search and download functionality here
        query = st.text_input("Enter your research query")
        
        if st.button("Search and Download Papers"):
            if query:
                # Simulate paper download
                st.success("Papers downloaded successfully!")
                # In a real implementation, you would call your existing paper download function
                # and store the chat_id for later use
                st.session_state.chat_id = "example_chat_id_123"
                
                # Set the processed data folder path
                downloads_dir = Path("downloads")
                st.session_state.processed_data_folder = str(downloads_dir / st.session_state.chat_id / "processed_data")
                
                st.info(f"Chat ID: {st.session_state.chat_id}")
                st.info(f"Processed data will be stored in: {st.session_state.processed_data_folder}")
            else:
                st.error("Please enter a query")
    
    with tab2:
        st.header("Process Downloaded PDFs")
        
        if st.session_state.chat_id:
            if st.button("Process PDFs"):
                # Simulate PDF processing
                st.success("PDFs processed successfully!")
                # In a real implementation, you would call your PDF processing function
                # Example: process_pdfs(chat_id=st.session_state.chat_id)
        else:
            st.warning("Please search and download papers first")
    
    with tab3:
        st.header("Index Documents in Pinecone")
        
        # Setup Pinecone button in sidebar
        if st.sidebar.button("Setup Pinecone"):
            setup_pinecone()  # This function is added by integrate_pinecone_workflow
        
        if st.session_state.processed_data_folder and st.session_state.chat_id:
            if st.button("Index Documents"):
                # Index the documents using the function added by integrate_pinecone_workflow
                index_name = index_documents(
                    processed_data_folder=st.session_state.processed_data_folder,
                    chat_uuid=st.session_state.chat_id
                )
                
                st.success(f"Documents indexed successfully in Pinecone index: {index_name}")
        else:
            st.warning("Please process PDFs first")
    
    with tab4:
        st.header("Query Pinecone Index")
        
        if st.session_state.pinecone_index_name:
            search_query = st.text_input("Enter your search query")
            top_k = st.slider("Number of results", min_value=1, max_value=20, value=5)
            
            if st.button("Search"):
                if search_query:
                    # Query the index using the function added by integrate_pinecone_workflow
                    query_documents(query=search_query, top_k=top_k)
                else:
                    st.error("Please enter a search query")
        else:
            st.warning("Please index documents in Pinecone first")

if __name__ == "__main__":
    main() 