requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain
langchain-google-genai
google-generativeai
pinecone
openai
tqdm>=4.65.0

# PDF processing
PyMuPDF>=1.22.0
nltk>=3.8.1

# Async support
aiohttp>=3.8.5
asyncio>=3.4.3

# Data handling
numpy>=1.24.3
pandas>=2.0.3

# Research paper downloader dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
streamlit>=1.24.0
pyperclip>=1.8.2
semanticscholar>=0.4.1
httpx>=0.25.0

# Optional dependencies for advanced features
# Uncomment if needed
# langchain>=0.1.0
# google-generativeai>=0.3.0

# Added from the code block
pinecone-client>=2.2.0
uuid>=1.30
argparse>=1.4.0 